cloud-computing/newsapiii/wheelify-api/requirements.txt
Flask
tensorflow
Flask
tensorflow
Pillow
requests
opencv-python
scikit-learn
scikit-image
matplotlib
pandas
google-cloud-storage
google-cloud-firestore